% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm-package.R
\docType{data}
\name{attendance}
\alias{attendance}
\title{Attendance data}
\format{The data are a matrix with 54 rows and 51 columns containing a 2 if the female was seen at the rookery and 1 if not seen.}
\description{
Observations of 54 female sea lions with pups over 51 consecutive days.
}
\keyword{datasets}

