% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_emission_exponentials.R
\name{BetaModel}
\alias{BetaModel}
\title{Beta Emission Model}
\description{
Models emission probabilities for variables bounded in \link{0, 1} using the Beta distribution.
}
\section{Super class}{
\code{hsmm::EmissionModel} -> \code{BetaModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alpha}}{Shape parameter α for each state.}

\item{\code{beta}}{Shape parameter β for each state.}

\item{\code{density}}{Matrix of emission probabilities for each observation and state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BetaModel-new}{\code{BetaModel$new()}}
\item \href{#method-BetaModel-compute_density}{\code{BetaModel$compute_density()}}
\item \href{#method-BetaModel-update}{\code{BetaModel$update()}}
\item \href{#method-BetaModel-clone}{\code{BetaModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetaModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-BetaModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaModel$new(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetaModel-compute_density"></a>}}
\if{latex}{\out{\hypertarget{method-BetaModel-compute_density}{}}}
\subsection{Method \code{compute_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaModel$compute_density(params, Pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetaModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-BetaModel-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaModel$update(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetaModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BetaModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
