% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitbeta.R
\name{dlogitbeta}
\alias{dlogitbeta}
\alias{rlogitbeta}
\alias{rlogitbeta}
\title{Logit-Beta density function}
\usage{
dlogitbeta(x, a, b, log = FALSE)

rlogitbeta(n, a, b)
}
\arguments{
\item{x}{vector of quantiles}

\item{a}{first shape parameter of a beta distribution}

\item{b}{second shape parameter of a beta distribution}

\item{n}{number of observations}
}
\value{
density or random variable generation
}
\description{
Calculates the PDF of the logit-beta distribution. The logit-beta distribution is defined as the 
logit transformation of a beta distributed random variable
}
\author{
Devin S. Johnson
}
