% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwell_time_distributions.R
\name{dt.values}
\alias{dt.values}
\title{Compute dwell time distribution function values for HSMM models}
\usage{
dt.values(dtf, n, theta)
}
\arguments{
\item{dtf}{dwell time distribution function}

\item{n}{number of discrete states for dwell time}

\item{theta}{parameter vector for dwell time distributions}
}
\value{
vector of probabilities for dwell times from approximation using geometric tail
}
\description{
Compute dwell time distribution function values for HSMM models
}

