% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_factory.R
\name{DurationModelFactory}
\alias{DurationModelFactory}
\title{Factory for Duration Models}
\description{
Factory for Duration Models

Factory for Duration Models
}
\details{
Creates and returns a duration model depending on whether covariates are provided.
If no covariates are present, a geometric duration model is used; otherwise, a
Gompertz-based duration model is chosen.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DurationModelFactory-create}{\code{DurationModelFactory$create()}}
\item \href{#method-DurationModelFactory-clone}{\code{DurationModelFactory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DurationModelFactory-create"></a>}}
\if{latex}{\out{\hypertarget{method-DurationModelFactory-create}{}}}
\subsection{Method \code{create()}}{
Constructs a duration model instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationModelFactory$create(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{List of parameters including possible \code{covariates_q}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object inheriting from \code{DurationModel}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DurationModelFactory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DurationModelFactory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationModelFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
