% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms_em.R
\name{em.algorithm}
\alias{em.algorithm}
\title{EM Algorithm for Hidden Semi-Markov Models (HSMM)}
\usage{
em.algorithm(params, emission.model, transition.model, duration.model)
}
\arguments{
\item{params}{A list containing model parameters and settings, including:
\itemize{
\item{\code{n_states}}{Number of hidden states.}
\item{\code{max_dwell}}{Maximum dwell time (duration) per state.}
\item{\code{max_iter}}{Maximum number of EM iterations.}
\item{\code{tol}}{Convergence tolerance for log-likelihood.}
\item{\code{verbose}}{Logical, print progress if TRUE.}
}}

\item{emission.model}{A list of functions defining the emission model,
including:
\itemize{
\item{\code{density}}{Matrix of emission densities for each observation and state.}
\item{\code{update}}{Function to update emission parameters.}
}}

\item{transition.model}{A list defining the transition dynamics,
with elements:
\itemize{
\item{\code{gamma}}{3D array of transition probabilities (time x from x to).}
\item{\code{Pi}}{Initial state probability vector.}
\item{\code{compute_omega, compute_gamma, update}}{Functions to update transition parameters.}
}}

\item{duration.model}{A list of duration-related functions:
\itemize{
\item{\code{compute_beta, compute_p.array}}{Functions to update dwell-time distributions.}
\item{\code{p.array}}{Array of duration probabilities.}
}}
}
\value{
A numeric vector of log-likelihood values over all EM iterations.
}
\description{
Performs the full Expectation-Maximization (EM) algorithm for an HSMM.
Iteratively calls \code{em.step()} until convergence or until the
maximum number of iterations is reached.
}
\details{
The algorithm tracks the log-likelihood (\code{llk}) at each iteration and
terminates when the absolute difference between successive values is less than
\code{params$tol}. If \code{params$verbose = TRUE}, progress is printed at each iteration.
}
\seealso{
\code{\link{em.step}}, \code{\link{backward.forward}}
}
