% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_factory.R
\name{EmissionModelFactory}
\alias{EmissionModelFactory}
\title{Factory for Emission Models}
\description{
Factory for Emission Models

Factory for Emission Models
}
\details{
Creates and returns an emission model instance based on the specified
\code{model_type} in the \code{params} list. Supports a variety of
emission families such as Gaussian, Poisson, Gamma, Beta, Torus, etc.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmissionModelFactory-create}{\code{EmissionModelFactory$create()}}
\item \href{#method-EmissionModelFactory-clone}{\code{EmissionModelFactory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionModelFactory-create"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionModelFactory-create}{}}}
\subsection{Method \code{create()}}{
Constructs an emission model instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionModelFactory$create(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{List of model parameters including \code{model_type}.}

\item{\code{Pi}}{Numeric vector of initial state probabilities.}

\item{\code{post.pi}}{Matrix of posterior probabilities per state.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object inheriting from \code{EmissionModel}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionModelFactory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionModelFactory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionModelFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
