% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_emission_exponentials.R
\name{ExponentialModel}
\alias{ExponentialModel}
\title{Exponential Emission Model}
\description{
Models continuous non-negative data using an exponential distribution.
}
\section{Super class}{
\code{hsmm::EmissionModel} -> \code{ExponentialModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lambda}}{Numeric vector of rate parameters.}

\item{\code{density}}{Matrix of emission probabilities for each observation and state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExponentialModel-new}{\code{ExponentialModel$new()}}
\item \href{#method-ExponentialModel-compute_density}{\code{ExponentialModel$compute_density()}}
\item \href{#method-ExponentialModel-update}{\code{ExponentialModel$update()}}
\item \href{#method-ExponentialModel-clone}{\code{ExponentialModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExponentialModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExponentialModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialModel$new(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExponentialModel-compute_density"></a>}}
\if{latex}{\out{\hypertarget{method-ExponentialModel-compute_density}{}}}
\subsection{Method \code{compute_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialModel$compute_density(params, Pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExponentialModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-ExponentialModel-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialModel$update(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExponentialModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExponentialModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
