% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_attendance.R
\name{fit_attendance}
\alias{fit_attendance}
\title{Fitting code for sea lion attendance data using hsmm}
\usage{
fit_attendance(df, ddl, dtf, pformula, omega, initial = NULL,
  method = "nlminb", debug = FALSE)
}
\arguments{
\item{df}{encounter history matrix where each row is sequence for each occasion (column)}

\item{ddl}{design data list of dwell time dataframe (dt) and sighting probability dataframe (p)}

\item{dtf}{list of length s where s is the number of states. The list elements specify the states in order
and contain a dwell time function (fct), a formula for the dwell time function parameter (at present only a
single parameter is allowed), and the number of time steps for the distribution (steps)}

\item{pformula}{formula to be applied to ddl$p to create the design matrix for the sighting probability model}

\item{omega}{an s by s matrix with known transition values between the states}

\item{initial}{vector of parameter initial values; must equal number of parameters if provided}

\item{method}{method to be used with optim; currently only a single method is supported}

\item{debug}{if TRUE will print out parameter values and negative log-likelihood value at current parameter values}
}
\value{
fitted list of model results
}
\description{
Computes MLEs of parameters for dwell time distributions and sighting probabilities to data in df, based on
model defined by ddl (design data list), dtf (dwell time distributions), pformula (formula for sighting probabilities),
and omega (known state transition matrix).
}
\examples{
{ 
data(attendance)
attendance=attendance[attendance$year==1999,]
# 4 state model - pre-birth, birth period, at sea, on land
omega=matrix(c(0,1,0,0,0,0,1,0,0,0,0,1,0,0,1,0),byrow=TRUE,ncol=4)
# strata.labels only used to create values in design data; values in encounter history are just 
# 0 (not seen) or 1 (seen) and not strata values;  encounter history is converted to 1 (not seen) and 2 (seen)
# process data
xp=process.data(attendance,model="ATTEND",strata.labels=c("P","B","S","L"))
# make design data and then delete fields that aren't used including fix which has default settings that
# are not useful for this model
ddl=make.design.data(xp)
# fixed real values
ddl$p$fix=NULL
ddl$dt$fix=NULL
ddl$dt$cohort=NULL
ddl$dt$Cohort=NULL
ddl$dt$age=NULL
ddl$dt$Age=NULL
ddl$p$cohort=NULL
ddl$p$Cohort=NULL
ddl$p$age=NULL
ddl$p$Age=NULL
# set p=0 for "S" at sea 
ddl$p$fix=ifelse(ddl$p$stratum\%in\%c("P","L","B"),NA,0)
# set p=0 for occasions in which not a single animal was seen; presumably no effort
ddl$p$fix[ddl$p$time\%in\%which(colSums(xp$ehmat)==nrow(xp$ehmat))]=0
# create covariate for birth and on-land p
ddl$p$birth=ifelse(ddl$p$stratum=="P",0,1)
# specify shifted poisson distributions for each state with different numbers of steps
state1=list(fct=shifted_poisson,steps=25,formula=~1)
state2=list(fct=shifted_poisson,steps=12,formula=~1)
state3=list(fct=shifted_poisson,steps=6,formula=~period)
state4=list(fct=shifted_poisson,steps=3,formula=~period)
dtf=list(state1,state2,state3,state4)
model<-fit_attendance(xp$ehmat,ddl=ddl,dtf=dtf,pformula=~birth,omega=omega,debug=TRUE)
par(mfrow=c(2,2))
plot_dt(model,range=15,labels=c("pre-birth","birth","at sea","on land"))
}
}
\author{
Jeff Laake
}
\references{
Zucchini, W. and I.L. MacDonald. 2009. Hidden Markov Models for Time Series: An Introduction using R. Chapman and Hall, Boca Raton, FL. See page 82.
}
\keyword{utility}

