% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia_funzione.R
\name{fit_HMM}
\alias{fit_HMM}
\title{Fit di un HMM personalizzato}
\usage{
fit_HMM(Omega, p_array, lfit, Pi, ld, d, K, M = 0)
}
\arguments{
\item{Omega}{matrice KxK di transizione}

\item{p_array}{matrice (n x sum(ld)) di probabilità}

\item{lfit}{matrice (n x sum(ld)) di likelihoods log}

\item{Pi}{distribuzione iniziale (lunghezza sum(ld))}

\item{ld}{vettore delle lunghezze dei blocchi}

\item{d}{etichette dei blocchi}

\item{K}{numero di blocchi}

\item{M}{parametro extra (se serve, altrimenti 0)}
}
\value{
lista con log-likelihood, posteriori, ecc.
}
\description{
Fit di un HMM personalizzato
}
