% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm2hmm.r
\name{gamma_dtd}
\alias{gamma_dtd}
\title{Computes dwell time distribution values and transition matrix}
\usage{
gamma_dtd(pars, dtf, mv, omega)
}
\arguments{
\item{pars}{list of parameter vectors; length of list is m - number of states}

\item{dtf}{list of length m with a function for each state}

\item{mv}{vector with number of steps (m_i) for each dwell time distribution; length of mv is m - the number of states}

\item{omega}{known m by m state transition matrix}
}
\value{
expanded state transition matrix defining HSMM
}
\description{
Create expanded transition matrix for HMM from aggregated state transition
matrix (omega) and dwell time distributions (dtf) calling hsmm2hmm.
}

