% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_emission_exponentials.R
\name{GammaModel}
\alias{GammaModel}
\title{Gamma Emission Model}
\description{
Models emission probabilities using a Gamma distribution.
}
\section{Super class}{
\code{hsmm::EmissionModel} -> \code{GammaModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{shape}}{Numeric vector of shape parameters.}

\item{\code{scale}}{Numeric vector of scale parameters.}

\item{\code{density}}{Matrix of emission probabilities for each observation and state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GammaModel-new}{\code{GammaModel$new()}}
\item \href{#method-GammaModel-compute_density}{\code{GammaModel$compute_density()}}
\item \href{#method-GammaModel-update}{\code{GammaModel$update()}}
\item \href{#method-GammaModel-clone}{\code{GammaModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GammaModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-GammaModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModel$new(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GammaModel-compute_density"></a>}}
\if{latex}{\out{\hypertarget{method-GammaModel-compute_density}{}}}
\subsection{Method \code{compute_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModel$compute_density(params, Pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GammaModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-GammaModel-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModel$update(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GammaModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GammaModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
