% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_emission_exponentials.R
\name{GaussianModel}
\alias{GaussianModel}
\title{Gaussian Emission Model for HMM/HSMM}
\description{
Gaussian Emission Model for HMM/HSMM

Gaussian Emission Model for HMM/HSMM
}
\details{
Models the emission distribution of each hidden state as a multivariate
Gaussian. Used within an HMM or HSMM framework.
}
\section{Super class}{
\code{hsmm::EmissionModel} -> \code{GaussianModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mu}}{Matrix (n_states x n_dim) of mean vectors.}

\item{\code{sigma}}{3D array (n_dim x n_dim x n_states) of covariance matrices.}

\item{\code{density}}{Matrix of computed emission densities for each observation and state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GaussianModel-new}{\code{GaussianModel$new()}}
\item \href{#method-GaussianModel-compute_density}{\code{GaussianModel$compute_density()}}
\item \href{#method-GaussianModel-update}{\code{GaussianModel$update()}}
\item \href{#method-GaussianModel-clone}{\code{GaussianModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianModel-new}{}}}
\subsection{Method \code{new()}}{
Initializes Gaussian model parameters (\code{mu}, \code{sigma}) using weighted sample estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianModel$new(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{List of model parameters including \code{data} and \code{n_states}.}

\item{\code{Pi}}{Initial state probabilities.}

\item{\code{post.pi}}{Posterior probabilities per state (from previous iteration).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianModel-compute_density"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianModel-compute_density}{}}}
\subsection{Method \code{compute_density()}}{
Computes Gaussian emission densities for all observations and states.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianModel$compute_density(params, Pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianModel-update}{}}}
\subsection{Method \code{update()}}{
Updates the Gaussian model parameters (\code{mu}, \code{sigma}) using new posterior weights.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianModel$update(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
