% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwell_time_distributions.R
\name{geometric}
\alias{geometric}
\title{Pre-defined dwell time distributions for HSMM models}
\usage{
geometric(theta, n)
}
\arguments{
\item{theta}{parameter vector for dwell time distributions}

\item{n}{number of discrete states for dwell time}

\item{dtf}{dwell time distribution list}

\item{range}{single value or vector of values with length = number of states; range value used from 1:range to plot distribution}

\item{fitted}{hsmm model object object}

\item{dm}{vector for combining parameter values for dt distribution}

\item{labels}{labels to be used for states in plotting}
}
\value{
vector of probabilities for well times
}
\description{
Various discrete dwell time distributions defined in Zucchini et al second edition.
Note: if you use unstructured without a geometric tail (gt), the number of dwell times must exceed the longest 
observed sequence in the state or -lnl could become undefined (Nan).
}
\author{
Jeff Laake
}
\references{
Zucchini, W., MacDonald, I.L. and Langrock, R. 2016. Hidden Markov Models for Time Series: 
An introduction using R, 2nd ed. CRC Press.
}
\keyword{utility}

