% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_likelihood.R
\name{hsmm_likelihood}
\alias{hsmm_likelihood}
\title{Hidden semi-Markov likelihood function Pre-defined dwell time distributions for HSMM models}
\usage{
hsmm_likelihood(pars, type, data, ddl, dtf, fct_gamma, fct_dmat, fct_delta,
  pformula, omega, debug = FALSE, mat = FALSE)
}
\arguments{
\item{pars}{parameter vector}

\item{type}{vector with number of parameters in each of the mv dwell time distributions followed by number of parameters in state dependent observation matrix (dmat) and initial vector (delta)}

\item{data}{matrix of observations; row is a sequence; column is an occasion; value is 1 to k where k is number of possible observations which match rows of dmat}

\item{ddl}{design data list of dwell time dataframe (dt) and sighting probability dataframe (p)}

\item{dtf}{list of length m with a function for each state}

\item{fct_gamma}{function that specifies the ttansition matrix (gamma)}

\item{fct_dmat}{function that specifies the state dependent observation matrix (dmat)}

\item{fct_delta}{function that specifies the initial probability vector}

\item{pformula}{formula to be applied to ddl$p to create the design matrix for the sighting probability model}

\item{omega}{known m by m state transition matrix}

\item{debug}{if TRUE, prints the parameter values and -log likelihood}

\item{mat}{if TRUE, return matrices rather than -log likelihood}
}
\value{
negative log-likelihood value or list of matrices depending on value of mat
}
\description{
Various discrete dwell time distributions defined in Zucchini et al second edition
}
\references{
Zucchini, W., MacDonald, I.L. and Langrock, R. 2016. Hidden Markov Models for Time Series: 
An introduction using R, 2nd ed. CRC Press.
}
\author{
Jeff Laake
}
\keyword{utility}
