% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwell_time_distributions.R
\name{plot_dt}
\alias{plot_dt}
\title{Plot dwell time distributions for HSMM models}
\usage{
plot_dt(object, range, dm = NULL, labels = NULL, ...)
}
\arguments{
\item{object}{fitted hsmm model object object}

\item{range}{single value or vector of values with length = number of states; range value used from 1:range to plot distribution}

\item{dm}{vector for combining parameter values for dt distribution}

\item{labels}{labels to be used for states in plotting}

\item{...}{plot parameters}
}
\value{
none
}
\description{
Plot dwell time distributions for HSMM models
}

