% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_emission_exponentials.R
\name{PoissonModel}
\alias{PoissonModel}
\title{Poisson Emission Model}
\description{
Models emission probabilities using a Poisson distribution for count data.
}
\section{Super class}{
\code{hsmm::EmissionModel} -> \code{PoissonModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lambda}}{Numeric vector of Poisson rate parameters (\eqn{\lambda_k}).}

\item{\code{density}}{Matrix of emission probabilities for each observation and state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PoissonModel-new}{\code{PoissonModel$new()}}
\item \href{#method-PoissonModel-compute_density}{\code{PoissonModel$compute_density()}}
\item \href{#method-PoissonModel-update}{\code{PoissonModel$update()}}
\item \href{#method-PoissonModel-clone}{\code{PoissonModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PoissonModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-PoissonModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonModel$new(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PoissonModel-compute_density"></a>}}
\if{latex}{\out{\hypertarget{method-PoissonModel-compute_density}{}}}
\subsection{Method \code{compute_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonModel$compute_density(params, Pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PoissonModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-PoissonModel-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonModel$update(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PoissonModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PoissonModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
