% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_emission_exponentials.R
\name{StudentTModel}
\alias{StudentTModel}
\title{Student-t Emission Model}
\description{
Models emission probabilities using a Student-t distribution, allowing for
heavier tails compared to the Gaussian model.
}
\section{Super class}{
\code{hsmm::EmissionModel} -> \code{StudentTModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mu}}{Mean for each state.}

\item{\code{sigma}}{Scale (standard deviation) for each state.}

\item{\code{df}}{Degrees of freedom.}

\item{\code{density}}{Matrix of emission probabilities for each observation and state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudentTModel-new}{\code{StudentTModel$new()}}
\item \href{#method-StudentTModel-compute_density}{\code{StudentTModel$compute_density()}}
\item \href{#method-StudentTModel-update}{\code{StudentTModel$update()}}
\item \href{#method-StudentTModel-clone}{\code{StudentTModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudentTModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudentTModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudentTModel$new(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudentTModel-compute_density"></a>}}
\if{latex}{\out{\hypertarget{method-StudentTModel-compute_density}{}}}
\subsection{Method \code{compute_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudentTModel$compute_density(params, Pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudentTModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-StudentTModel-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudentTModel$update(params, Pi, post.pi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudentTModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudentTModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudentTModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
