% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_factory.R
\name{TransitionModelFactory}
\alias{TransitionModelFactory}
\title{Factory for Transition Models}
\description{
Factory for Transition Models

Factory for Transition Models
}
\details{
Creates and returns a transition model depending on the presence of covariates.
Without covariates, a homogeneous transition model is used; with covariates,
a time-varying transition model is constructed.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransitionModelFactory-create}{\code{TransitionModelFactory$create()}}
\item \href{#method-TransitionModelFactory-clone}{\code{TransitionModelFactory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionModelFactory-create"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionModelFactory-create}{}}}
\subsection{Method \code{create()}}{
Constructs a transition model instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionModelFactory$create(params, p.array)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{List of parameters, including \code{covariates_omega}.}

\item{\code{p.array}}{Duration probability array (used by the transition model).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object inheriting from \code{TransitionModel}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionModelFactory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionModelFactory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionModelFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
