% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwell_time_distributions.R
\name{dt.values}
\alias{dt.values}
\title{Compute dwell time distribution function values for HSMM models}
\usage{
dt.values(dtf, n, theta)
}
\arguments{
\item{dtf}{dwell time distribution function}

\item{n}{number of discrete states for dwell time}

\item{theta}{parameter vector for dwell time distributions}
}
\value{
vector of probabilities for dwell times from approximation using geometric tail
}
\description{
Compute dwell time distribution function values for HSMM models
}
\examples{
par(mfrow=c(2,1))
plot(1:10,shifted_poisson(log(2),10),ylab="Probability mass d(r)",
   xlab="Time steps (r)",main="",type="b",pch=1,yaxt="n")
x=dt.values(dtf=list(fct=shifted_poisson,steps=2),10,theta=log(2))
lines(x,type="b",lty=2,pch=2)
plot(1:10,shifted_poisson(log(2),10),ylab="Probability mass d(r)",
  xlab="Time steps (r)",main="",type="b",pch=1,yaxt="n")
x=dt.values(dtf=list(fct=shifted_poisson,steps=3),10,theta=log(2))
lines(x,type="b",lty=2,pch=2)

}

